modificaciones y seguimientos

+++ basic/controllers/EtiquetasController.php
@@ -7,6 +7,8 @@ use app\models\Etiquetas;
 use app\models\EtiquetasSearch;
 use app\models\Usuarios;
 use app\models\UsuarioAvisos;
+use app\models\ActividadEtiquetas;
+use app\models\ActividadEtiquetasQuery;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
@@ -21,19 +23,10 @@ class EtiquetasController extends Controller
      */
     public function behaviors()
     {
-				if( Yii::$app->user->isGuest ){
+		if( Yii::$app->user->isGuest ){
 			$this->redirect(Yii::$app->request->baseURL."\site\login");
-		}else{		
-			if(Yii::$app->user->identity->username!="admin"){
-				$u=Usuarios::findOne(Yii::$app->user->identity->id);
-				if($u->rol!='A'){
-					$this->redirect(Yii::$app->request->baseURL."\site\login");
-				}
-			}
-		}
-
-		
-        return [
+		}					
+		return [
             'verbs' => [
                 'class' => VerbFilter::className(),
                 'actions' => [
@@ -56,24 +49,21 @@ class EtiquetasController extends Controller
 			$msg=$_SESSION['msg'];
 			unset($_SESSION['msg']);
 		}
+		$rol='A';
+		if(Yii::$app->user->identity){
+			$u=Usuarios::findOne(Yii::$app->user->identity->id);
+			if($u){
+				$rol=$u->rol;
+			}
+		}
         return $this->render('index', [
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
-			'msg'=>$msg
+			'msg'=>$msg,
+			'rol'=>$rol,
         ]);
     }
 
-    /**
-     * Displays a single Etiquetas model.
-     * @param string $id
-     * @return mixed
-     */
-    public function actionView($id)
-    {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
-    }
 
     /**
      * Creates a new Etiquetas model.
@@ -88,15 +78,17 @@ class EtiquetasController extends Controller
 			$aviso=new UsuarioAvisos();
 			$aviso->fecha=date("Y/m/d");
 			$aviso->clase_aviso_id='N';
-			$aviso->texto="Nueva etiqueta creada ".$model->nombre.". URL: ".Yii::$app->request->baseURL."\etiquetas\view?id=".$model->id;
+			$aviso->texto="Nueva etiqueta creada ".$model->id." ".$model->nombre;
 			$aviso->destino_usuario_id=0;
 			$aviso->origen_usuario_id=0;
-			$u=Usuarios::findOne(Yii::$app->user->identity->id);
-			if($u){
-				$aviso->origen_usuario_id=$u->id;
+			if(Yii::$app->user->identity){
+				$u=Usuarios::findOne(Yii::$app->user->identity->id);
+				if($u){
+					$aviso->origen_usuario_id=$u->id;
+				}
 			}
 			$aviso->save();
-            return $this->redirect(['view', 'id' => $model->id]);
+             return $this->redirect(['index']);
         } else {
             return $this->render('create', [
                 'model' => $model,
@@ -112,15 +104,52 @@ class EtiquetasController extends Controller
      */
     public function actionUpdate($id)
     {
+		if(Yii::$app->user->identity){
+			$u=Usuarios::findOne(Yii::$app->user->identity->id);
+			if($u){
+				if($u->rol!='A') $this->redirect(Yii::$app->request->baseURL."\site\login");
+			}
+		}
+		
         $model = $this->findModel($id);
-
+		
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            return $this->redirect(['index']);
         } else {
+			$etiquetas = Etiquetas::find()->all();
+			$todas = array();
+			foreach($etiquetas as $e){
+				if($e->id!=$id)
+					$todas[$e->id]=$e->nombre;
+			}
             return $this->render('update', [
                 'model' => $model,
+				'todas' => $todas
             ]);
         }
+    }
+	public function actionUnifica($id)
+    {
+		
+		if(Yii::$app->user->identity){
+			$u=Usuarios::findOne(Yii::$app->user->identity->id);
+			if($u){
+				if($u->rol!='A') $this->redirect(Yii::$app->request->baseURL."\site\login");
+			}
+		}
+        $searchModel = new EtiquetasSearch();
+        $dataProvider = $searchModel->search([]);
+		$relaciones=ActividadEtiquetas::find()->all();
+		foreach($relaciones as $relacion){
+			if($relacion->etiqueta_id==Yii::$app->request->queryParams['iduni']){
+				$relacion->etiqueta_id=Yii::$app->request->queryParams['id'];
+				$relacion->save();
+			}
+		}
+		$this->findModel(Yii::$app->request->queryParams['iduni'])->delete();
+		$msg="etiquetas unificadas";
+		
+        return $this->redirect(['index']);
     }
 
     /**
@@ -131,16 +160,28 @@ class EtiquetasController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
-		/*$todos=Actividades::find()->all();
-		foreach($todos as $c){
-			if($c->clasificacion_id==$id  || $c->clasificacion_etiqueta_id==$id){
-				$_SESSION['msg']="No se puede eliminar, la clasificación tiene etiquetas relacionadas";
-				return $this->redirect(['index']);
+		if(Yii::$app->user->identity){
+			$u=Usuarios::findOne(Yii::$app->user->identity->id);
+			if($u){
+				if($u->rol!='A') $this->redirect(Yii::$app->request->baseURL."\site\login");
 			}
-		}*/
-    
-        return $this->redirect(['index']);
+		}
+		$todos=ActividadEtiquetas::find()->all();
+		$sepuede=true;
+		foreach($todos as $c){
+			if($c->etiqueta_id==$id){
+				$sepuede=false;
+				
+			}
+		}
+		if($sepuede){
+			$this->findModel($id)->delete();
+			return $this->redirect(['index']);
+		}else{
+			$_SESSION['msg']="No se puede eliminar, la etiqueta tiene actividades relacionadas";
+				return $this->redirect(['index']);
+		}
+   
     }
 
     /**
